improve dev_env_init.sh

+++ scripts/dev_env_init.sh
@@ -8,7 +8,11 @@ while [ -h "$SOURCE" ]; do # resolve $SOURCE until the file is no longer a symli
 done
 DIR="$( cd -P "$( dirname "$SOURCE" )/.." >/dev/null 2>&1 && pwd )"
 cd $DIR
-echo "creating default environment variables"
-cp ./scripts/.env.template ./.env
+
+if [! -f .env ]
+then
+    echo "creating default environment variables"
+    cp ./scripts/.env.template ./.env
+fi
 echo "install pip dependencies"
 pipenv install
